Add `client` and `server ` features
